Moved handler function into it's own file
