+++ src/mcp-index.ts
@@ -9,14 +9,12 @@
 import { Context, Hono } from 'hono';
 import { createMiddleware } from 'hono/factory';
 import { cors } from 'hono/cors';
-import { isInitializeRequest } from '@modelcontextprotocol/sdk/types';
-import { RESPONSE_ALREADY_SENT } from '@hono/node-server/utils/response';
 
 import { ServerConfig } from './types/mcp';
-import { MCPSession, TransportType } from './services/mcpSession';
-import { SSEServerTransport } from '@modelcontextprotocol/sdk/server/sse';
+import { MCPSession } from './services/mcpSession';
 import { SessionStore } from './services/sessionStore';
 import { createLogger } from './utils/logger';
+import { handleMCPRequest, handleSSEMessages } from './handlers/mcpHandler';
 
 const logger = createLogger('MCP-Gateway');
 
@@ -114,317 +112,7 @@ app.get('/', (c) => {
  * Main MCP endpoint with transport detection
  */
 app.all('/:serverId/mcp', hydrateContext, sessionMiddleware, async (c) => {
-  logger.debug(`${c.req.method} ${c.req.url}`, { headers: c.req.raw.headers });
-  const serverConfig = c.var.serverConfig;
-  let session = c.var.session;
-
-  // Detect transport type from headers
-  const acceptHeader = c.req.header('Accept');
-  const isSSERequest =
-    c.req.method === 'GET' && acceptHeader?.includes('text/event-stream');
-
-  // Parse body for POST requests
-  const body = c.req.method === 'POST' ? await c.req.json() : null;
-  logger.debug(`Body: ${body ? JSON.stringify(body, null, 2) : 'null'}`);
-
-  // Check if this is an initialization request
-  if (body && isInitializeRequest(body)) {
-    // Determine client transport type from request
-    // For now, assume POST with initialize = streamable-http
-    // Real SSE clients would establish the event stream first
-    const clientTransportType: TransportType = 'streamable-http';
-
-    logger.debug(
-      `Initialize request - defaulting to streamable-http transport`
-    );
-
-    // Create new session if needed
-    if (!session) {
-      logger.info(
-        `Creating new session for server: ${c.req.param('serverId')}`
-      );
-
-      // Normal new session creation
-      session = new MCPSession(serverConfig);
-      sessionStore.set(session.id, session);
-    }
-
-    // Initialize or restore the session
-    logger.debug(
-      `Session ${session.id}: Client requesting ${clientTransportType} transport`
-    );
-
-    try {
-      // Use the new initializeOrRestore method
-      await session.initializeOrRestore(clientTransportType);
-
-      const capabilities = session.getTransportCapabilities();
-      logger.info(
-        `Session ${session.id}: Transport established ${capabilities?.clientTransport} -> ${capabilities?.upstreamTransport}`
-      );
-    } catch (error) {
-      logger.error(`Failed to initialize session ${session.id}`, error);
-      sessionStore.delete(session.id);
-      session = undefined;
-    }
-
-    // Handle the request if session is valid
-    if (!session) {
-      logger.error('Failed to create or restore session');
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32000,
-            message: 'Failed to initialize session',
-          },
-          id: (body as any)?.id || null,
-        },
-        500
-      );
-    }
-
-    // Handle the request
-    const { incoming: req, outgoing: res } = c.env as any;
-    logger.debug(`Session ${session.id}: Handling initialize request`);
-
-    // Set session ID header for client
-    if (res && res.setHeader) {
-      res.setHeader('mcp-session-id', session.id);
-    }
-
-    await session.handleRequest(req, res, body);
-    logger.debug(`Session ${session.id}: Initialize request completed`);
-
-    return RESPONSE_ALREADY_SENT;
-  }
-
-  // Handle GET requests for established sessions
-  // Both SSE and Streamable HTTP use GET requests with event-stream accept headers
-  if (c.req.method === 'GET' && session) {
-    // Get the transport type that was determined during initialization
-    const clientTransportType = session.getClientTransportType();
-
-    if (!clientTransportType) {
-      logger.error(`Session ${session.id} has no transport type set`);
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32000,
-            message: 'Session not properly initialized',
-          },
-          id: null,
-        },
-        500
-      );
-    }
-
-    // Ensure session is active or can be restored
-    try {
-      await session.initializeOrRestore(clientTransportType);
-      logger.debug(
-        `Session ${session.id} ready for ${clientTransportType} connection`
-      );
-    } catch (error) {
-      logger.error(`Failed to prepare session ${session.id}`, error);
-      sessionStore.delete(session.id);
-
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32000,
-            message: 'Failed to restore session. Please reinitialize.',
-          },
-          id: null,
-        },
-        500
-      );
-    }
-
-    const { incoming: req, outgoing: res } = c.env as any;
-
-    // Route based on the actual transport type, not the accept header
-    if (clientTransportType === 'sse') {
-      // For true SSE clients, set up the SSE stream
-      res.writeHead(200, {
-        'Content-Type': 'text/event-stream',
-        'Cache-Control': 'no-cache, no-transform',
-        Connection: 'keep-alive',
-        'X-Session-Id': session.id,
-      });
-
-      // Handle connection cleanup on close/error
-      const currentSession = session;
-      res.on('close', () => {
-        logger.info(`SSE connection closed for session ${currentSession.id}`);
-        sessionStore.delete(currentSession.id);
-        currentSession
-          .close()
-          .catch((err) => logger.error('Error closing session', err));
-      });
-
-      res.on('error', (error: any) => {
-        logger.error(
-          `SSE connection error for session ${currentSession.id}`,
-          error
-        );
-        sessionStore.delete(currentSession.id);
-        currentSession
-          .close()
-          .catch((err) => logger.error('Error closing session', err));
-      });
-
-      // Initialize the SSE transport with the response object
-      const transport = session.initializeSSETransport(res);
-
-      // Start the SSE transport
-      await transport.start();
-
-      return RESPONSE_ALREADY_SENT;
-    } else {
-      // For Streamable HTTP clients, let the transport handle the request
-      await session.handleRequest(req, res);
-      return RESPONSE_ALREADY_SENT;
-    }
-  }
-
-  // For non-initialization requests, require session
-  if (!session) {
-    // For GET requests without session, check if it's a true SSE client
-    // True SSE clients will have ONLY text/event-stream in Accept header
-    const isPureSSE =
-      c.req.method === 'GET' && acceptHeader === 'text/event-stream';
-
-    if (isPureSSE) {
-      logger.info('Creating new session for pure SSE client');
-      session = new MCPSession(serverConfig);
-      sessionStore.set(session.id, session);
-      c.set('session', session);
-
-      // Initialize the session with SSE transport
-      try {
-        await session.initializeOrRestore('sse');
-        logger.info(`SSE session ${session.id} initialized`);
-      } catch (error) {
-        logger.error(`Failed to initialize SSE session ${session.id}`, error);
-        sessionStore.delete(session.id);
-        return c.json(
-          {
-            jsonrpc: '2.0',
-            error: {
-              code: -32000,
-              message: 'Failed to initialize SSE session',
-            },
-            id: null,
-          },
-          500
-        );
-      }
-    } else {
-      logger.warn(
-        `No session found - method: ${c.req.method}, sessionId: ${c.req.header('mcp-session-id')}`
-      );
-      if (c.req.method === 'POST') {
-        return c.json(
-          {
-            jsonrpc: '2.0',
-            error: {
-              code: -32000,
-              message: 'Session required. Please initialize first.',
-            },
-            id: null,
-          },
-          400
-        );
-      } else {
-        return c.json(
-          {
-            jsonrpc: '2.0',
-            error: {
-              code: -32001,
-              message: 'Session not found',
-            },
-            id: null,
-          },
-          404
-        );
-      }
-    }
-  }
-
-  // Ensure session is properly initialized before handling request
-  if (session && !isInitializeRequest(body)) {
-    try {
-      // Determine transport type from request or use saved capabilities
-      const acceptHeader = c.req.header('Accept');
-      const isCurrentSSERequest =
-        c.req.method === 'GET' && acceptHeader?.includes('text/event-stream');
-      const detectedTransportType: TransportType = isCurrentSSERequest
-        ? 'sse'
-        : 'streamable-http';
-
-      // Use detected transport type or fall back to saved capabilities
-      const transportType =
-        session.getTransportCapabilities()?.clientTransport ||
-        detectedTransportType;
-
-      // This will handle all states (dormant, active, initializing) appropriately
-      await session.initializeOrRestore(transportType);
-      logger.debug(`Session ${session.id} ready for request handling`);
-    } catch (error) {
-      logger.error(`Failed to prepare session ${session.id}`, error);
-      sessionStore.delete(session.id);
-
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32000,
-            message: 'Failed to restore session. Please reinitialize.',
-          },
-          id: (body as any)?.id || null,
-        },
-        500
-      );
-    }
-  }
-
-  // Handle request through session
-  const { incoming: req, outgoing: res } = c.env as any;
-
-  try {
-    logger.debug(`Session ${session.id}: Handling ${c.req.method} request`);
-    await session.handleRequest(req, res, body);
-  } catch (error: any) {
-    logger.error(`Error handling request for session ${session.id}`, error);
-
-    // If this is a session initialization error, try to clean up and respond
-    if (error?.message?.includes('Session not initialized')) {
-      logger.error(
-        `CRITICAL: Session ${session.id} initialization failed unexpectedly`
-      );
-      sessionStore.delete(session.id);
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32000,
-            message:
-              'Session initialization failed in handleRequest. Please reconnect.',
-          },
-          id: body?.id || null,
-        },
-        500
-      );
-    }
-
-    // Re-throw other errors
-    throw error;
-  }
-
-  return RESPONSE_ALREADY_SENT;
+  return handleMCPRequest(c, sessionStore);
 });
 
 /**
@@ -447,77 +135,7 @@ app.post(
   hydrateContext,
   sessionMiddleware,
   async (c) => {
-    logger.debug(`POST ${c.req.url}`);
-    const sessionId = c.req.query('sessionId');
-
-    if (!sessionId) {
-      logger.warn('POST /messages: Missing session ID in query');
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32000,
-            message: 'Session ID required in query parameter',
-          },
-          id: null,
-        },
-        400
-      );
-    }
-
-    const session = sessionStore.get(sessionId);
-    if (!session) {
-      logger.warn(`POST /messages: Session ${sessionId} not found`);
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32001,
-            message: 'Session not found',
-          },
-          id: null,
-        },
-        404
-      );
-    }
-
-    // Ensure session is ready for SSE messages
-    try {
-      const transportType = session.getClientTransportType() || 'sse';
-      await session.initializeOrRestore(transportType);
-      logger.debug(`Session ${sessionId} ready for SSE messages`);
-    } catch (error) {
-      logger.error(
-        `Failed to prepare session ${sessionId} for SSE messages`,
-        error
-      );
-      sessionStore.delete(sessionId);
-
-      return c.json(
-        {
-          jsonrpc: '2.0',
-          error: {
-            code: -32001,
-            message:
-              'Failed to restore session during SSE reconnection. Please reinitialize.',
-          },
-          id: null,
-        },
-        500
-      );
-    }
-
-    const body = await c.req.json();
-    logger.debug(`Session ${sessionId}: Processing ${body.method} message`);
-
-    // Access the underlying Node.js request/response
-    const { incoming: req, outgoing: res } = c.env as any;
-
-    // Handle the message through the SSE transport
-    const transport = session.getDownstreamTransport() as SSEServerTransport;
-    await transport.handlePostMessage(req, res, body);
-
-    return RESPONSE_ALREADY_SENT;
+    return handleSSEMessages(c, sessionStore);
   }
 );
