+++ wrangler.toml
@@ -6,6 +6,7 @@ compatibility_flags = [ "nodejs_compat" ]
 [vars]
 ENVIRONMENT = 'dev'
 CUSTOM_HEADERS_TO_IGNORE = []
+LOG_LEVEL = 'debug'
 
 #
 #Configuration for DEVELOPMENT environment
@@ -16,6 +17,7 @@ name = "rubeus-dev"
 [env.staging.vars]
 ENVIRONMENT = 'staging'
 CUSTOM_HEADERS_TO_IGNORE = []
+LOG_LEVEL = 'info'
 
 
 #
@@ -28,3 +30,4 @@ logpush=true
 [env.production.vars]
 ENVIRONMENT = 'production'
 CUSTOM_HEADERS_TO_IGNORE = []
+LOG_LEVEL = 'error'
